+++ dbms/src/Flash/Coprocessor/RemoteRequest.h
@@ -34,7 +34,10 @@ using DAGSchema = std::vector<DAGColumnInfo>;
 
 struct RemoteRequest
 {
-    RemoteRequest(tipb::DAGRequest && dag_request_, DAGSchema && schema_, std::vector<pingcap::coprocessor::KeyRange> && key_ranges_)
+    RemoteRequest(
+        tipb::DAGRequest && dag_request_,
+        DAGSchema && schema_,
+        std::vector<pingcap::coprocessor::KeyRange> && key_ranges_)
         : dag_request(std::move(dag_request_))
         , schema(std::move(schema_))
         , key_ranges(std::move(key_ranges_))
@@ -43,6 +46,13 @@ struct RemoteRequest
     DAGSchema schema;
     /// the sorted key ranges
     std::vector<pingcap::coprocessor::KeyRange> key_ranges;
-    static RemoteRequest build(const RegionRetryList & retry_regions, DAGContext & dag_context, const TiDBTableScan & table_scan, const TiDB::TableInfo & table_info, const tipb::Executor * selection, LoggerPtr & log);
+    static RemoteRequest build(
+        const RegionRetryList & retry_regions,
+        DAGContext & dag_context,
+        const TiDBTableScan & table_scan,
+        const TiDB::TableInfo & table_info,
+        const String & sel_executor_id,
+        const std::vector<const tipb::Expr *> & sel_conditions,
+        const LoggerPtr & log);
 };
 } // namespace DB
